+++ CHANGELOG.md
@@ -11,7 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added new story to reload data with sample code.
 -   Implemented className and style properties to the table and the actions.
 -   Added reload tests.
+
+### Changed
 -   Changed max-lines allowed in DataTableInterfaces.ts.
+-   Changed max-lines allowed in DataTableRouted.tsx.
 
 ## [2.0.3] - 2022-04-19