+++ CHANGELOG.md
@@ -6,13 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.1.0] - 2022-05-02
+
 ### Added
+
 -   Dynamic table handler to reload the data from an external source.
 -   Added new story to reload data with sample code.
 -   Implemented className and style properties to the table and the actions.
 -   Added reload tests.
 
 ### Changed
+
 -   Changed max-lines allowed in DataTableInterfaces.ts.
 -   Changed max-lines allowed in DataTableRouted.tsx.
 
@@ -40,7 +45,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   created package :tada: 
 
-[Unreleased]: https://github.com/neolution-ch/react-data-table/compare/2.0.3...HEAD
+[Unreleased]: https://github.com/neolution-ch/react-data-table/compare/2.1.0...HEAD
+
+[2.1.0]: https://github.com/neolution-ch/react-data-table/compare/2.0.3...2.1.0
 
 [2.0.3]: https://github.com/neolution-ch/react-data-table/compare/2.0.2...2.0.3